+++ Runtime/WrapConfigData.cs
@@ -4,14 +4,41 @@ using UnityEngine;
 namespace SOSXR.ConfigData
 {
     /// <summary>
-    ///     The only purpose of this class is to hold a reference to a ConfigData in the Hierarchy.
-    ///     This can be useful for visualising the ConfigData.
-    ///     It can be important in the (unlikely) case that the ConfigData is not referenced by any other object, and therefore
-    ///     not loaded into memory. However, if your ConfigData is not referenced by any other object, then you're clearly not
-    ///     using it, and you should probably not worry about it.
+    ///     This is a wrapper for the ConfigData to ensure that the ConfigData is initialized and destroyed correctly.
+    ///     The init and destroy are there to ensure the Json file is updated when the relevant values change.
     /// </summary>
     public class WrapConfigData : MonoBehaviour
     {
-        public BaseConfigData ConfigData;
+        [HideInInspector] public TestConfigData ConfigData;
+
+
+        private void Awake()
+        {
+            ConfigData.Initialize();
+        }
+
+
+        private void OnEnable()
+        {
+            ConfigData.Subscribe(nameof(ConfigData.ShowAffordances), obj => Lemmeno());
+        }
+
+
+        private void Lemmeno()
+        {
+            Debug.LogFormat(this, "ShowAffordances changed to {0}", ConfigData.ShowAffordances);
+        }
+
+
+        private void OnDisable()
+        {
+            ConfigData.Unsubscribe(nameof(ConfigData.ShowAffordances), obj => Lemmeno());
+        }
+
+
+        private void OnDestroy()
+        {
+            ConfigData.Uninitialize();
+        }
     }
 }